+++ spinta/cli/config.py
@@ -1,3 +1,4 @@
+from spinta.components import Context, Store
 from typing import List
 from typing import Optional
 
@@ -6,6 +7,7 @@ from typer import Context as TyperContext
 from typer import Option
 from typer import echo
 
+from spinta import commands
 from spinta.cli.helpers.manifest import convert_str_to_manifest_path
 from spinta.cli.helpers.store import prepare_manifest
 from spinta.core.enums import Mode
@@ -19,7 +21,7 @@ def config(
     fmt: KeyFormat = KeyFormat.cfg,
 ):
     """Show current configuration values"""
-    context = configure_context(ctx.obj)
+    context: Context = configure_context(ctx.obj)
     rc = context.get("rc")
     rc.dump(*name, fmt=fmt)
 
@@ -32,10 +34,16 @@ def check(
 ):
     """Check configuration and manifests"""
     manifests = convert_str_to_manifest_path(manifests)
-    context = configure_context(ctx.obj, manifests, mode=mode, check_names=check_names)
-    context.set("strict_check", True)
+    context: Context = configure_context(ctx.obj, manifests, mode=mode, check_names=check_names)
     prepare_manifest(context, ensure_config_dir=True, full_load=True)
     manager = context.get("error_manager")
+
+    store: Store = context.get("store")
+    if store.manifest:
+        for model in commands.get_models(context, store.manifest).values():
+            if model.external and model.external.resource and model.external.resource.backend:
+                commands.check(context, model, model.external.resource.backend)
+
     handler = manager.handler
 
     if handler.get_counts():