+++ spinta/manifests/tabular/helpers.py
@@ -52,6 +52,7 @@ from spinta.exceptions import (
     NestedDataTypeMismatch,
     NoModelDefined,
     PropertyNotFound,
+    SourceOrPrepareNotAllowed,
 )
 from spinta.manifests.components import Manifest
 from spinta.manifests.helpers import load_manifest_nodes
@@ -687,6 +688,10 @@ class PropertyReader(TabularReader):
     def _parse_prepare(self):
         if "prepare" in self.data:
             self.data["prepare"] = _parse_spyna(self, self.data["prepare"])
+            prepare = self.data["prepare"]
+            if prepare is not NA and isinstance(prepare, dict) and prepare.get("name") == "getattr":
+                if external_name := self.data.get("external", {}).get("name"):
+                    raise SourceOrPrepareNotAllowed(source=external_name)
 
     def _append_prepare(self, row: Dict[str, str], prepare: str):
         if "prepare" in self.data: